requirements.txt
Faker==8.10.1
selenium
PyHamcrest
pip
webdriver-manager
requests
busypie
pytest
allure-pytest
selene==2.0.0b5
pytest-tldr